Update mataa_measure_impedance.m
Add latency to the arguments of mataa_measure_impedance, so that it is passed to mataa_measure_signal_response. Actually, it is currently expected by mataa_measure_impedance, but never received simply because it is not one of the function arguments —and that causes a warning message to appear to the user.